Add go.mod and update .gitignore
This commit introduces a go.mod file, setting up the necessary package dependencies for our mediadracker application. The dependencies used in the project are willingly updated to the latest versions to leverage bug fixes and performance enhancements offered by them.

In addition, this commit updates the gitignore file with a line (.idea),