Refactor mediadata package and update Movie struct
This commit involves a major refactoring of the mediadata package for overall improved readability and maintainability. The package tmdb.go is renamed to movie.go, indicating a more specific function tied to the code. Temporary code related to TV show functionality has been removed as it is not yet applicable and was creating unnecessary complexity.

The Movie structure was extended to include fields such as Rating and RatingCount to provide a more detailed presentation of individual movies. Such enhancement will improve the user experience by providing more comprehensive information about the movies. The improvements also reflect in the MovieDetails, TvShow and TvShowDetails structures that now include additional fields such as Genres, Cast and Studio.

Changes are also reflected in the main.go file, demonstrating the use of the updated MovieClient interface. Specifically, it now includes an implementation for retrieving details of a chosen movie (rating, genre, etc.).

These changes are made to create a more solid, informative and user-friendly platform for tracking movie information.