Refactor tmdb config to use API provided cast details
The buildCast function is refactored to buildMovieCast in the tmdb config file. Earlier, it was casting the fetched details internally. With recent TMDB API changes, these details are now provided in a better format by the API itself. Hence, the casting is no longer needed, and we can directly use these details.